requirements.txt
albumentations==1.0.3
certifi==2020.6.20
cycler==0.11.0
dataclasses==0.8
decorator==4.4.2
imageio==2.8.0
importlib-metadata==4.8.2
joblib==1.1.0
kiwisolver==1.3.1
matplotlib==3.3.4
networkx==2.5.1
numpy==1.17.0
opencv-python-headless==4.5.4.60
Pillow==8.4.0
prefetch-generator==1.0.1
prettytable==2.4.0
protobuf==3.19.1
pyparsing==3.0.6
python-dateutil==2.8.2
PyWavelets==1.1.1
PyYAML==6.0
scikit-image==0.17.2
scikit-learn==0.24.2
scipy==1.5.4
six==1.16.0
sklearn==0.0
tensorboardX==2.4.1
threadpoolctl==3.0.0
tifffile==2020.9.3
torch==1.8.0+cu111
torchaudio==0.8.0
torchsummary==1.5.1
torchvision==0.9.0+cu111
tqdm==4.45.0
typing_extensions==4.0.0
wcwidth==0.2.5
yacs==0.1.8
zipp==3.6.0
